Rename `heap` to `Liballoc`
Moved so that the new heap code can live in Heap.cpp/.hpp
