+++ Kernel/Devices/Graphics/graphics.cpp
@@ -16,7 +16,7 @@
 #include <Devices/Graphics/graphics.hpp>
 #include <stddef.h>
 #include <stdint.h>
-#include <Memory/heap.hpp>
+#include <Memory/LibAlloc.hpp>
 #include <Memory/paging.hpp>
 #include <Bootloader/Handoff.hpp>
 #include <Library/stdio.hpp>
